fix(streaming): correct frame compaction and byte decode safety
Problem:
WebSocket frame consumption used incorrect compaction math, which could
move more bytes than remain in the receive buffer. Byte assembly for
32-bit fields also used left-shifts on promoted signed values.

Solution:
Use a single bytes_consumed value when compacting the WS receive buffer
and subtract that exact length from the remaining buffer. Cast each
input byte to std::uint32_t before shifting when decoding extended
payload sizes and metadata encoding fields.

Notes:
This keeps protocol behavior unchanged while removing memory-safety and
undefined-behavior risks in untrusted input paths.

+++ src/detail/streaming_protocol.cpp
@@ -41,9 +41,17 @@ wss::detail::streaming_protocol::decode_header(
             return header;
 
         if (use_tcp_protocol)
-            header.payload_size = (data[0] << 24) | (data[1] << 16) | (data[2] << 8) | data[3];
+            header.payload_size =
+                (static_cast<std::uint32_t>(data[0]) << 24)
+                | (static_cast<std::uint32_t>(data[1]) << 16)
+                | (static_cast<std::uint32_t>(data[2]) << 8)
+                | static_cast<std::uint32_t>(data[3]);
         else
-            header.payload_size = (data[3] << 24) | (data[2] << 16) | (data[1] << 8) | data[0];
+            header.payload_size =
+                (static_cast<std::uint32_t>(data[3]) << 24)
+                | (static_cast<std::uint32_t>(data[2]) << 16)
+                | (static_cast<std::uint32_t>(data[1]) << 8)
+                | static_cast<std::uint32_t>(data[0]);
 
         data += sizeof(std::uint32_t);
         size -= sizeof(std::uint32_t);
